requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

numpy==1.26.4
faiss-cpu==1.7.4

torch==2.0.1+cpu
sentence-transformers==2.2.2
huggingface-hub==0.16.4
transformers==4.30.2

fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.4
requests
beautifulsoup4
lxml
scikit-learn
pandas
